+++ ui/litellm-dashboard/src/app/(dashboard)/agents/_components/AgentsPanel.tsx
