+++ ui/litellm-dashboard/src/app/(dashboard)/guardrails/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import GuardrailsPanel from "./_components";
+import GuardrailsPanel from "./_components/GuardrailsPanel";
 import useAuthorized from "@/app/(dashboard)/hooks/useAuthorized";
 
 export default function Guardrails() {
